template_content/playground/hello_world/helloworld.py
import beaker
import pyteal as pt

app = beaker.Application("HelloWorldApp")


@app.external
def hello(name: pt.abi.String, *, output: pt.abi.String) -> pt.Expr:
    return output.set(pt.Concat(pt.Bytes("Hello, "), name.get()))
