template_content/pyproject.toml
[tool.poetry]
name = "playground"
version = "0.1.0"
description = "Interactive examples of Algorand examples built with beaker"
authors = ["MakerX <cto@makerx.com.au>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "1.0.0b2"
algokit-utils = "0.1.3"
python-dotenv = "^1.0.0"
